rename utils.external to utils.matrix_calibration and remove old utils.freight
